Remove old picker and replace everywhere with new one
