+++ front/app/components/admin/DatePickers/DateSinglePicker/DateSinglePicker.stories.tsx
@@ -1,10 +1,6 @@
 import React, { useState } from 'react';
 
-import { Box } from '@citizenlab/cl2-component-library';
-
-import DateSinglePicker from './index2';
-
-import OldDateSinglePicker from '.';
+import DateSinglePicker from '.';
 
 import type { Meta, StoryObj } from '@storybook/react';
 
@@ -38,27 +34,5 @@ export const Disabled: Story = {
   args: {
     disabled: true,
     onChange: () => {},
-  },
-};
-
-// TODO remove
-const WrapperOld = () => {
-  const [selectedDate, setSelectedDate] = useState<Date>();
-
-  return (
-    <Box w="120px">
-      <OldDateSinglePicker
-        selectedDate={selectedDate ?? null}
-        onChange={(date) => {
-          setSelectedDate(date ?? undefined);
-        }}
-      />
-    </Box>
-  );
-};
-
-export const Old = {
-  render: () => {
-    return <WrapperOld />;
   },
 };
