+++ front/app/modules/commercial/smart_groups/components/UserFilterConditions/ValueSelector/DateValueSelector.tsx
@@ -18,7 +18,7 @@ const DateValueSelector = ({ value, onChange }: Props) => {
 
   return (
     <DateSinglePicker
-      selectedDate={typeof value === 'string' ? new Date(value) : null}
+      selectedDate={typeof value === 'string' ? new Date(value) : undefined}
       onChange={handleOnChange}
     />
   );
